backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
pytesseract>=0.3.10
numpy>=1.26.4
opencv-python>=4.10.0.84
Pillow>=10.3.0
openai-whisper
pydub
setuptools
